media: dvb-core: Fix use-after-free due on race condition at dvb_net
jira ROC-3067
cve CVE-2022-45886
commit 4172385b0c9ac366dcab78eda48c26814b87ed1a

A race condition may occur between the .disconnect function, which
is called when the device is disconnected, and the dvb_device_open()
function, which is called when the device node is open()ed.
This results in several types of UAFs.

The root cause of this is that you use the dvb_device_open() function,
which does not implement a conditional statement
that checks 'dvbnet->exit'.

So, add 'remove_mutex` to protect 'dvbnet->exit' and use
locked_dvb_net_open() function to check 'dvbnet->exit'.

[mchehab: fix a checkpatch warning]

    Link: https://lore.kernel.org/linux-media/20221117045925.14297-3-imv4bel@gmail.com
    Signed-off-by: Hyunwoo Kim <imv4bel@gmail.com>
    Signed-off-by: Mauro Carvalho Chehab <mchehab@kernel.org>
    (cherry picked from commit 4172385b0c9ac366dcab78eda48c26814b87ed1a)
Signed-off-by: Ronnie Sahlberg <rsahlberg@ciq.com>