Add fips logging changes to alg_test_hash() -> __alg_test_hash() -> test_hash_vec() -> test_hash_vec_cfg() -> test_ahash_vec_cfg().
Diff versus FIPS8.
In fips8 we call the hash-function three different times for each vector
with different alignment.
To reduce logging and testing fips-8 only performs logging in one case,
when align offset == 0.
In fips-9 the hash test functions are structured differently and we no
longer test different alignment but instead we test two different
api's to use each hash function.
Similarly to fips-8 we here will only perform the functional test for
one of the API, test_ahash... and not test_shash... again because we
want to just test the algorithm itself and not the api used to access
the allgorithm.

Since the loop over the vectors is several stack frames removed from the
actual test in fips-9 we can not abort logging errors like fips-8 did.
So, we will log every failed vector for every algorithm instead of just
once per algorithm.
If fips-8 behavior is important we can add an argument we pass between
the two sites.

Signed-off-by: Ronnie Sahlberg <rsahlberg@ciq.com>