perf: Fix perf_event_validate_size()
jira SECO-54
cve CVE-2023-6931
commit 382c27f4ed28f803b1f1473ac2d8db0afc795a1b
upstream-diff trivial diff against upstream as we do not have
PERF_FORMAT_LOST(119a784c8) in 8.8. PERF_FORMAT_LOST is a counter
that was exposed to userspace post 8.8 so we do not need it backported.

Budimir noted that perf_event_validate_size() only checks the size of
the newly added event, even though the sizes of all existing events
can also change due to not all events having the same read_format.

When we attach the new event, perf_group_attach(), we do re-compute
the size for all events.

Fixes: a723968c0ed3 ("perf: Fix u16 overflows")
    Reported-by: Budimir Markovic <markovicbudimir@gmail.com>
    Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
    (cherry picked from commit 382c27f4ed28f803b1f1473ac2d8db0afc795a1b)
Signed-off-by: Ronnie Sahlberg <rsahlberg@ciq.com>