drm/vmwgfx: Fix invalid reads in fence signaled events
jira VULN-5435
cve CVE-2024-36960
commit-author Zack Rusin <zack.rusin@broadcom.com>
commit a37ef7613c00f2d72c8fc08bd83fb6cc76926c8c

Correctly set the length of the drm_event to the size of the structure
that's actually used.

The length of the drm_event was set to the parent structure instead of
to the drm_vmw_event_fence which is supposed to be read. drm_read
uses the length parameter to copy the event to the user space thus
resuling in oob reads.

	Signed-off-by: Zack Rusin <zack.rusin@broadcom.com>
Fixes: 8b7de6aa8468 ("vmwgfx: Rework fence event action")
	Reported-by: zdi-disclosures@trendmicro.com # ZDI-CAN-23566
	Cc: David Airlie <airlied@gmail.com>
        CC: Daniel Vetter <daniel@ffwll.ch>
	Cc: Zack Rusin <zack.rusin@broadcom.com>
	Cc: Broadcom internal kernel review list <bcm-kernel-feedback-list@broadcom.com>
	Cc: dri-devel@lists.freedesktop.org
	Cc: linux-kernel@vger.kernel.org
	Cc: <stable@vger.kernel.org> # v3.4+
	Reviewed-by: Maaz Mombasawala <maaz.mombasawala@broadcom.com>
	Reviewed-by: Martin Krastev <martin.krastev@broadcom.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240425192748.1761522-1-zack.rusin@broadcom.com
(cherry picked from commit a37ef7613c00f2d72c8fc08bd83fb6cc76926c8c)
	Signed-off-by: Greg Rose <g.v.rose@ciq.com>