netfilter: nf_tables: defer gc run if previous batch is still pending
jira VULN-597
subsystem-sync netfilter:nf_tables 4.18.0-534
commit-author Florian Westphal <fw@strlen.de>
commit 8e51830e29e12670b4c10df070a4ea4c9593e961

Don't queue more gc work, else we may queue the same elements multiple
times.

If an element is flagged as dead, this can mean that either the previous
gc request was invalidated/discarded by a transaction or that the previous
request is still pending in the system work queue.

The latter will happen if the gc interval is set to a very low value,
e.g. 1ms, and system work queue is backlogged.

The sets refcount is 1 if no previous gc requeusts are queued, so add
a helper for this and skip gc run if old requests are pending.

Add a helper for this and skip the gc run in this case.

Fixes: f6c383b8c31a ("netfilter: nf_tables: adapt set backend to use GC transaction API")
	Signed-off-by: Florian Westphal <fw@strlen.de>
	Reviewed-by: Pablo Neira Ayuso <pablo@netfilter.org>
(cherry picked from commit 8e51830e29e12670b4c10df070a4ea4c9593e961)
	Signed-off-by: Greg Rose <g.v.rose@ciq.com>